requirements.txt
appdirs==1.4.4
casttube==0.2.1
certifi==2020.4.5.1
chardet==3.0.4
fontawesome==5.10.1.post1
geojson==2.5.0
idna==2.9
ifaddr==0.1.6
Pillow==7.1.2
protobuf==3.12.0
PyChromecast==5.2.0
pyglet==1.5.5
pyowm==2.10.0
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
urllib3==1.25.9
zeroconf==0.26.1
